pyproject.toml
[project]
name = "claude-gsuite-admin-mcp"
version = "0.1.0"
description = "Claude CLI Google Workspace Admin MCP Server - Comprehensive Google Workspace administration tools for Claude CLI"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Ryan Robson", email = "ryan@robworks.info"}
]
keywords = ["mcp", "claude", "google-workspace", "gsuite", "admin", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Systems Administration",
    "Topic :: Communications :: Email",
    "Topic :: Office/Business :: Office Suites",
]

dependencies = [
    "beautifulsoup4>=4.12.3",
    "google-api-python-client>=2.154.0",
    "google-auth>=2.23.3",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.1",
    "httplib2>=0.22.0",
    "mcp>=1.3.0",
    "oauth2client==4.1.3",
    "python-dotenv>=1.0.1",
    "pytz>=2024.2",
    "requests>=2.32.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pyright>=1.1.389",
]

[project.urls]
Homepage = "https://github.com/ryanrobson/claude-gsuite-admin-mcp"
Repository = "https://github.com/ryanrobson/claude-gsuite-admin-mcp"
Issues = "https://github.com/ryanrobson/claude-gsuite-admin-mcp/issues"
Documentation = "https://github.com/ryanrobson/claude-gsuite-admin-mcp/blob/main/docs/"

[project.scripts]
claude-gsuite-admin-mcp = "claude_gsuite_admin:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/claude_gsuite_admin"]

[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "oauth2client.*",
    "googleapiclient.*",
    "google.*",
    "mcp.*"
]
ignore_missing_imports = true